requirements.txt
email_validator==2.1.1
fastapi==0.95.0
uvicorn==0.20.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.6
alembic==1.10.1
pydantic==1.10.5
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
boto3==1.26.50
